[#1] refactor: Authority 위치 변경
